move guard and delimitr to common dir
